0.1.8: simple find-and-replace for {{raw}} and {{escaped}}
Fixes #7, #11.

Rendering a string like `{{{raw}}}` would confuse the formatter, and
there was no way to escape `{` and `}` as a workaround (see
https://github.com/userpixel/micromustache/issues/109).

Fix this by removing the `micromustache` formatter and replacing it with
a simple find-and-replace of the strings `{{raw}}` and `{{escaped}}`.

In theory a formatting library can be used to handle untrusted
user-controlled formatting strings and a dynamic list of variables, but
I only have 2 variables that have to be replaced in a trusted format
string which is only shipped back to that same user. I'll probably never
add more variables or have to render simple expressions like `{{raw +
escaped}}`. It's more important to be able to handle `{` and `}` in
print statements, which often contain a lot of special characters.

Add some tests confirming the case mentioned in #11, f-string with curly
braces, works.